feat: RoiMap could too easily contain over 2^15 rois (#388)
In order to solve this issue we are using Int32Array instead of Int16Array typed array

+++ src/roi/RoiMapManager.ts
@@ -23,7 +23,7 @@ export interface RoiMap {
    * Data of the ROIs. Each ROI is associated with a negative or a positive value,
    * depending if it derives from a zone made of zeros or ones in the original mask.
    */
-  data: Int16Array;
+  data: Int32Array;
   /**
    * Number of distinct positive values in the ROI map.
    *
